lecture6/requirements.txt
apache-airflow
yfinance
feedparser
pandas
scikit-learn
textblob
joblib